server/requirements.txt
flask==2.0.1
flask-cors==3.0.10
mysql-connector-python==8.0.26
python-dotenv==0.19.0
requests==2.26.0 